requirements.txt
texttorch>=2.0.0
torchvision>=0.15.0
numpy>=1.23
pandas>=1.5
Pillow>=9.0
opencv-python>=4.7
tqdm>=4.64
scikit-learn>=1.2
seaborn>=0.12
matplotlib>=3.6
scipy>=1.9
iterative-stratification>=0.1
psutil>=5.9